finalProject/requirements.txt
django==3.2
singleton_decorator
djangorestframework==3.12.4